Make use of all Concept Types
The 'type' field of MedMentions often includes several assignments, and you seem to have made the choice of only keeping the first type. The change I'm proposing creates additional entries ('entities') so that all type assignments are available for training, etc..
Quickly checking the difference in number of entities after this change, I'm seeing about 9% additional entities (in training set).

Just getting started with MedMentions, don't know if there may be some reason to just keep the first type, but if there isn't, this change may be of interest.